+++ src/test/java/com/example/dvlacme/DataFactory.java
@@ -4,28 +4,28 @@ import com.example.dvlacme.domain.Record;
 
 public class DataFactory {
 
-    public static final Integer REFERENCE_1 = 163817;
+    public static final Long REFERENCE_1 = 163817l;
     public static final String ACCOUNT_NUMBER_1 = "NL46ABNA0625805417";
     public static final String DESCRIPTION_1 = "Candy for Erik de Vries";
     public static final Double START_BALANCE_1 = 68.95;
     public static final Double MUTATION_1 = 9.92;
     public static final Double  END_BALANCE_1 = 78.87;
 
-    public static final Integer REFERENCE_2 = 117747;
+    public static final Long REFERENCE_2 = 117747l;
     public static final String ACCOUNT_NUMBER_2 = "NL69ABNA0433647324";
     public static final String DESCRIPTION_2 = "Clothes for Willem Dekker";
     public static final Double START_BALANCE_2 = 5429d;
     public static final Double MUTATION_2 = -939d;
     public static final Double  END_BALANCE_2 = 6368d;
 
-    public static final Integer REFERENCE_3 = 153938;
+    public static final Long REFERENCE_3 = 153938l;
     public static final String ACCOUNT_NUMBER_3 = "NL32RABO0195610843";
     public static final String DESCRIPTION_3 = "Subscription for Jan King";
     public static final Double START_BALANCE_3 = 14.83;
     public static final Double MUTATION_3 = -33.47;
     public static final Double  END_BALANCE_3 = -18.64;
 
-    public static Record makeRecord(Integer reference, String accountNumber, String description, Double startBalance, Double mutation, Double endBalance) {
+    public static Record makeRecord(Long reference, String accountNumber, String description, Double startBalance, Double mutation, Double endBalance) {
         Record record = new Record();
         record.setReference(reference);
         record.setAccountNumber(accountNumber);
